chore(cleanup): move mdx files out of root
