+++ apps/site/components/Downloads/Release/ReleaseCodeBox.tsx
@@ -11,7 +11,6 @@ import CodeBox from '#site/components/Common/CodeBox';
 import LinkWithArrow from '#site/components/Common/LinkWithArrow';
 import Link from '#site/components/Link';
 import WithReleaseAlertBox from '#site/components/withReleaseAlertBox';
-import { createSval } from '#site/next.jsx.compiler.mjs';
 import {
   ReleaseContext,
   ReleasesContext,
@@ -19,13 +18,14 @@ import {
 import type { DownloadSnippet } from '#site/types/download';
 import type { ReleaseContextType } from '#site/types/release';
 import { INSTALL_METHODS } from '#site/util/download';
+import { createEvaluator } from '#site/util/evaluator';
 
 // Creates a minimal JavaScript interpreter for parsing the JavaScript code from the snippets
 // Note: that the code runs inside a sandboxed environment and cannot interact with any code outside of the sandbox
 // It also does not have access to any Global or Window objects, nor it can execute code on the end-user's browser
 // It also only allows a return statement for a string and it forces the return value to also be a string and only be used
 // by Shiki to render the highlighted syntax. Hence XSS attacks or JavaScript injections are not possible.
-const interpreter = createSval({}, 'script');
+const interpreter = createEvaluator({}, 'script');
 
 /**
  * Parses a snippet string using the interpreter with the given release context
