+++ apps/site/mdx/plugins.mjs
@@ -12,7 +12,7 @@ import remarkTableTitles from './util/table';
 /**
  * Provides all our Rehype Plugins that are used within MDX
  */
-export const REHYPE_PLUGINS = [
+export const rehypePlugins = [
   // Generates `id` attributes for headings (H1, ...)
   rehypeSlug,
   // Automatically add anchor links to headings (H1, ...)
@@ -25,7 +25,7 @@ export const REHYPE_PLUGINS = [
 /**
  * Provides all our Remark Plugins that are used within MDX
  */
-export const REMARK_PLUGINS = [
+export const remarkPlugins = [
   // Support GFM syntax to be used within Markdown
   remarkGfm,
   // Generates metadata regarding headings
