--- apps/site/next.mdx.use.client.mjs
@@ -1,77 +1,0 @@
-'use strict';
-
-import Blockquote from '@node-core/ui-components/Common/Blockquote';
-import MDXCodeTabs from '@node-core/ui-components/MDX/CodeTabs';
-
-import DownloadButton from './components/Downloads/DownloadButton';
-import BlogPostLink from './components/Downloads/Release/BlogPostLink';
-import ChangelogLink from './components/Downloads/Release/ChangelogLink';
-import ReleaseDownloadLink from './components/Downloads/Release/DownloadLink';
-import ReleaseInstallationMethodDropdown from './components/Downloads/Release/InstallationMethodDropdown';
-import ReleaseOperatingSystemDropdown from './components/Downloads/Release/OperatingSystemDropdown';
-import ReleasePackageManagerDropdown from './components/Downloads/Release/PackageManagerDropdown';
-import ReleasePlatformDropdown from './components/Downloads/Release/PlatformDropdown';
-import ReleasePrebuiltDownloadButtons from './components/Downloads/Release/PrebuiltDownloadButtons';
-import ReleaseCodeBox from './components/Downloads/Release/ReleaseCodeBox';
-import ReleaseVersionDropdown from './components/Downloads/Release/VersionDropdown';
-import Link from './components/Link';
-import MDXCodeBox from './components/MDX/CodeBox';
-import MDXImage from './components/MDX/Image';
-import WithReleaseSelect from './components/withReleaseSelect';
-import { ReleaseProvider } from './providers/releaseProvider';
-
-/**
- * A full list of React Components that we want to pass through to MDX
- *
- * @satisfies {import('mdx/types').MDXComponents}
- */
-export const clientMdxComponents = {
-  // Renders MDX CodeTabs
-  CodeTabs: MDXCodeTabs,
-  // Renders a Download Button
-  DownloadButton,
-  // Renders a stateless Release Select Component
-  WithReleaseSelect,
-  // Group of components that enable you to select versions for Node.js
-  // releases and download selected versions. Uses `releaseProvider` as a provider
-  Release: {
-    // Provides an individual Node.js Release Context for Downloads
-    Provider: ReleaseProvider,
-    // Renders a drop-down menu to select a version
-    VersionDropdown: ReleaseVersionDropdown,
-    // Renders a drop-down menu to select a platform
-    InstallationMethodDropdown: ReleaseInstallationMethodDropdown,
-    // Renders a drop-down menu to select a package manager
-    PackageManagerDropdown: ReleasePackageManagerDropdown,
-    // Renders a drop-down menu to select a bitness
-    PlatformDropdown: ReleasePlatformDropdown,
-    // Renders a drop-down menu to select an operating system
-    OperatingSystemDropdown: ReleaseOperatingSystemDropdown,
-    // Renders a Blog Post Link for the selected release
-    BlogPostLink,
-    // Renders a Download Button for the selected release
-    PrebuiltDownloadButtons: ReleasePrebuiltDownloadButtons,
-    // Renders a Release CodeBox
-    ReleaseCodeBox,
-    // Renders a Changelog Link Button
-    ChangelogLink,
-    // Renders a DownloadLink Button
-    DownloadLink: ReleaseDownloadLink,
-  },
-};
-
-/**
- * A full list of wired HTML elements into custom React Components
- *
- * @type {import('mdx/types').MDXComponents}
- */
-export const htmlComponents = {
-  // Renders a Link Component for `a` tags
-  a: Link,
-  // Renders a Blockquote Component for `blockquote` tags
-  blockquote: Blockquote,
-  // Renders a CodeBox Component for `pre` tags
-  pre: MDXCodeBox,
-  // Renders an Image Component for `img` tags
-  img: MDXImage,
-};
