+++ apps/site/util/evaluator.ts
@@ -1,0 +1,18 @@
+'use strict';
+
+import Sval from 'sval';
+
+export const createEvaluator = (
+  dependencies: Record<string, unknown> = {},
+  mode: 'module' | 'script' = 'module'
+) => {
+  const svalInterpreter = new Sval({
+    ecmaVer: 'latest',
+    sandBox: true,
+    sourceType: mode,
+  });
+
+  svalInterpreter.import(dependencies);
+
+  return svalInterpreter;
+};
